+++ service/pod/keyv2/error.go
@@ -1,0 +1,10 @@
+package keyv2
+
+import "github.com/giantswarm/microerror"
+
+var wrongTypeError = microerror.New("wrong type")
+
+// IsWrongTypeError asserts wrongTypeError.
+func IsWrongTypeError(err error) bool {
+	return microerror.Cause(err) == wrongTypeError
+}
